scraper/requirements.txt
-i https://pypi.org/simple
boto3==1.42.83; python_version >= '3.9'
botocore==1.42.83; python_version >= '3.9'
certifi==2026.2.25; python_version >= '3.7'
charset-normalizer==3.4.7; python_version >= '3.7'
idna==3.11; python_version >= '3.8'
jmespath==1.1.0; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.2.2; python_version >= '3.10'
requests==2.33.1; python_version >= '3.10'
s3transfer==0.16.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==2.6.3; python_version >= '3.9'
